Add files via upload
INTRODUCTION

Welcome to my project focused on sentiment analysis for a university's public reviews. In this project aims to explore and compare the effectiveness of both machine learning (ML) and non-ML based techniques in analyzing sentiments expressed by the public towards the university. By scraping data through a web browser and transforming it into a structured dataframe, and conduct a comprehensive analysis to extract valuable insights from the collected reviews.